add option for clearance in all directions
The existing implementation of clearance only checks for obstacles along one diagonal.  This works for the problem as defined in the reference material, but it is not suitable for situations where one needs to check for clearance in a radius around the walkable point.  To address that need, this PR adds an option to check for clearance in all directions.  This change is backwards-compatible and preserves the existing check as the default behavior.